% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_profile.R
\name{analyze_data}
\alias{analyze_data}
\title{Analyze Dataset Characteristics}
\usage{
analyze_data(data, max_unique_check = 10000)
}
\arguments{
\item{data}{A data frame to analyze}

\item{max_unique_check}{Maximum unique values to check for categorical columns (default: 10000)}
}
\value{
A list containing data profile information
}
\description{
Analyze Dataset Characteristics
}
