% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{calculate_test_metrics}
\alias{calculate_test_metrics}
\title{Calculate Test Metrics}
\usage{
calculate_test_metrics(predictions, target_col, task_type)
}
\arguments{
\item{predictions}{Predictions with actual values}

\item{target_col}{Target column name}

\item{task_type}{Task type}
}
\value{
A tibble of metrics
}
\description{
Calculate Test Metrics
}
\keyword{internal}
