% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumption_checks.R
\name{check_overfitting}
\alias{check_overfitting}
\title{Check for Overfitting}
\usage{
check_overfitting(model_results)
}
\arguments{
\item{model_results}{Output from train_model()}
}
\value{
Character string with overfitting check (markdown format)
}
\description{
Check for Overfitting
}
