% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{create_model_spec}
\alias{create_model_spec}
\title{Create Model Specification}
\usage{
create_model_spec(engine, task_type)
}
\arguments{
\item{engine}{Tidymodels engine name}

\item{task_type}{Task type}
}
\value{
A parsnip model specification
}
\description{
Create Model Specification
}
\keyword{internal}
