% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{create_prep_recipe}
\alias{create_prep_recipe}
\title{Generate Data Preparation Recipe}
\usage{
create_prep_recipe(data, target_col, issues)
}
\arguments{
\item{data}{A data frame}

\item{target_col}{Target variable name}

\item{issues}{Output from identify_issues()}
}
\value{
A tidymodels recipe object
}
\description{
Generate Data Preparation Recipe
}
