% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_detector.R
\name{detect_ml_task}
\alias{detect_ml_task}
\title{Detect Machine Learning Task Type}
\usage{
detect_ml_task(data, target_col)
}
\arguments{
\item{data}{A data frame}

\item{target_col}{Name of the target variable}
}
\value{
A list with task type and recommendations
}
\description{
Detect Machine Learning Task Type
}
