% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{explain_classification_results}
\alias{explain_classification_results}
\title{Explain Classification Results in Plain English}
\usage{
explain_classification_results(metrics, target)
}
\arguments{
\item{metrics}{Test metrics tibble}

\item{target}{Target variable name}
}
\value{
Character string explanation
}
\description{
Explain Classification Results in Plain English
}
\keyword{internal}
