% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{explain_model_results}
\alias{explain_model_results}
\title{Generate Plain English Model Summary}
\usage{
explain_model_results(model_results)
}
\arguments{
\item{model_results}{Output from train_model()}
}
\value{
A character string with plain English explanation
}
\description{
Generate Plain English Model Summary
}
