% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{extract_importance}
\alias{extract_importance}
\title{Extract Feature Importance}
\usage{
extract_importance(fitted_model, engine)
}
\arguments{
\item{fitted_model}{Fitted workflow}

\item{engine}{Model engine}
}
\value{
A tibble of feature importance or NULL
}
\description{
Extract Feature Importance
}
\keyword{internal}
