% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_recommender.R
\name{filter_compatible_methods}
\alias{filter_compatible_methods}
\title{Filter Methods by Task Compatibility}
\usage{
filter_compatible_methods(methods, task_type)
}
\arguments{
\item{methods}{Tibble of recommended methods}

\item{task_type}{"regression" or "classification"}
}
\value{
Filtered tibble of compatible methods
}
\description{
Removes methods that are not compatible with the task type
}
