% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{generate_prep_steps}
\alias{generate_prep_steps}
\title{Generate Human-Readable Prep Steps}
\usage{
generate_prep_steps(issues)
}
\arguments{
\item{issues}{Output from identify_issues()}
}
\value{
A character vector of preparation steps
}
\description{
Generate Human-Readable Prep Steps
}
