% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate Recommendations Based on Results}
\usage{
generate_recommendations(model_results)
}
\arguments{
\item{model_results}{Output from train_model()}
}
\value{
Character string with recommendations
}
\description{
Generate Recommendations Based on Results
}
\keyword{internal}
