% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{generate_tidymodels_code}
\alias{generate_tidymodels_code}
\title{Helper to generate tidymodels code}
\usage{
generate_tidymodels_code(method, task_info, data, profile)
}
\description{
Helper to generate tidymodels code
}
\keyword{internal}
