% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{get_method_display_name}
\alias{get_method_display_name}
\title{Get Method Display Name}
\usage{
get_method_display_name(engine)
}
\arguments{
\item{engine}{Engine name}
}
\value{
Human-readable name
}
\description{
Get Method Display Name
}
