% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{get_metric_color}
\alias{get_metric_color}
\title{Get Color for Metric Value}
\usage{
get_metric_color(value, metric_type)
}
\arguments{
\item{value}{Metric value}

\item{metric_type}{Type of metric}
}
\value{
Color string
}
\description{
Get Color for Metric Value
}
\keyword{internal}
