% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{has_tune_params}
\alias{has_tune_params}
\title{Check if Model Has Tuning Parameters}
\usage{
has_tune_params(spec)
}
\arguments{
\item{spec}{Model specification}
}
\value{
Logical
}
\description{
Check if Model Has Tuning Parameters
}
\keyword{internal}
