% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_profile.R
\name{identify_issues}
\alias{identify_issues}
\title{Identify Data Quality Issues}
\usage{
identify_issues(data, profile_data)
}
\arguments{
\item{data}{A data frame}

\item{profile_data}{Output from analyze_data()}
}
\value{
A tibble of identified issues
}
\description{
Identify Data Quality Issues
}
