% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_recommender.R
\name{is_model_compatible}
\alias{is_model_compatible}
\title{Check Model Compatibility with Task Type}
\usage{
is_model_compatible(engine, task_type)
}
\arguments{
\item{engine}{Tidymodels engine name}

\item{task_type}{"regression" or "classification"}
}
\value{
Logical indicating compatibility
}
\description{
Returns TRUE if the model engine is compatible with the task type
}
