% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_LearnTidyML}
\alias{launch_LearnTidyML}
\title{Launch LearnTidyML Shiny App}
\usage{
launch_LearnTidyML(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Logical; if TRUE (default), opens the app in the
system's default web browser. If FALSE, opens in RStudio viewer pane.}

\item{...}{Additional arguments passed to shiny::runApp}
}
\description{
Launch LearnTidyML Shiny App
}
