% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_detector.R
\name{needs_regularization}
\alias{needs_regularization}
\title{Detect if Regularization is Needed}
\usage{
needs_regularization(task_info)
}
\arguments{
\item{task_info}{Output from detect_ml_task()}
}
\value{
Logical indicating if regularization recommended
}
\description{
Detect if Regularization is Needed
}
