% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Create Confusion Matrix Visualization}
\usage{
plot_confusion_matrix(predictions, target_col)
}
\arguments{
\item{predictions}{Predictions tibble with .pred_class and actual}

\item{target_col}{Target column name}
}
\value{
A ggplot object
}
\description{
Create Confusion Matrix Visualization
}
