% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_importance}
\alias{plot_importance}
\title{Create Feature Importance Plot}
\usage{
plot_importance(importance, top_n = 15)
}
\arguments{
\item{importance}{Importance tibble}

\item{top_n}{Number of features to show (default: 15)}
}
\value{
A ggplot object
}
\description{
Create Feature Importance Plot
}
