% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_predictions_regression}
\alias{plot_predictions_regression}
\title{Create Predictions vs Actual Plot (Regression)}
\usage{
plot_predictions_regression(predictions, target_col)
}
\arguments{
\item{predictions}{Predictions tibble with .pred and actual columns}

\item{target_col}{Target column name}
}
\value{
A ggplot object
}
\description{
Create Predictions vs Actual Plot (Regression)
}
