% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Create Residual Plot (Regression)}
\usage{
plot_residuals(predictions, target_col)
}
\arguments{
\item{predictions}{Predictions tibble}

\item{target_col}{Target column name}
}
\value{
A ggplot object
}
\description{
Create Residual Plot (Regression)
}
