% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_recommender.R
\name{recommend_methods}
\alias{recommend_methods}
\title{Recommend ML Methods}
\usage{
recommend_methods(task_info, data_profile)
}
\arguments{
\item{task_info}{Output from detect_ml_task()}

\item{data_profile}{Output from analyze_data()}
}
\value{
A tibble of recommended methods with rationale
}
\description{
Recommend ML Methods
}
