% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_error_message}
\alias{safe_error_message}
\title{Safe Error Message Handler}
\usage{
safe_error_message(e, user_message = "An error occurred")
}
\arguments{
\item{e}{Error condition}

\item{user_message}{Generic user-facing message}
}
\value{
Character string with sanitized message
}
\description{
Sanitizes error messages to prevent information disclosure
}
\keyword{internal}
