% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_r_name}
\alias{safe_r_name}
\title{Safe Column Name for Code Generation}
\usage{
safe_r_name(name)
}
\arguments{
\item{name}{Column name}
}
\value{
Escaped column name
}
\description{
Escapes column names for safe use in generated R code
}
\keyword{internal}
