% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_dimensions}
\alias{validate_data_dimensions}
\title{Validate Data Dimensions}
\usage{
validate_data_dimensions(
  data,
  max_rows = 1e+05,
  max_cols = 1000,
  max_cells = 1e+07
)
}
\arguments{
\item{data}{A data frame}

\item{max_rows}{Maximum number of rows (default: 100000)}

\item{max_cols}{Maximum number of columns (default: 1000)}

\item{max_cells}{Maximum total cells (default: 10000000)}
}
\value{
TRUE if validation passes, throws error otherwise
}
\description{
Ensures data doesn't exceed resource limits
}
\keyword{internal}
