% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_target_column}
\alias{validate_target_column}
\title{Validate Target Column}
\usage{
validate_target_column(data, target_col)
}
\arguments{
\item{data}{A data frame}

\item{target_col}{Target column name}
}
\value{
TRUE if validation passes, throws error otherwise
}
\description{
Validates that a target column exists and is safe to use
}
\keyword{internal}
