% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_timeout}
\alias{with_timeout}
\title{Execute with Timeout}
\usage{
with_timeout(expr, timeout = 30)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{timeout}{Timeout in seconds (default: 30)}
}
\value{
Result of expression if successful
}
\description{
Wraps an expression with a timeout to prevent resource exhaustion
}
\keyword{internal}
