% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{create_default_recipe}
\alias{create_default_recipe}
\title{Create Default Recipe}
\usage{
create_default_recipe(data, target_col, task_type)
}
\arguments{
\item{data}{Training data}

\item{target_col}{Target column name}

\item{task_type}{Task type}
}
\value{
A recipe object
}
\description{
Create Default Recipe
}
\keyword{internal}
